trying to do delete user

+++ functions/index.js
@@ -1,5 +1,6 @@
 const functions = require('firebase-functions');
 const admin = require('firebase-admin');
+const { user } = require('firebase-functions/lib/providers/auth');
 // const { user } = require('firebase-functions/lib/providers/auth');
 // const { auth } = require('firebase');
 admin.initializeApp(functions.config().firebase);
@@ -44,28 +45,65 @@ exports.scrubUserFromDatabase = functions.auth.user()
                     if (userType === 'coach') {
                         deleteCoachData(userUID)
                             .then(() => {
+                                admin.firestore()
+                                    .collection('users')
+                                    .doc(user.uid)
+                                    .delete()
+                                    .then(() => {
+                                        return true
+                                    })
+                                    .catch(error => {
+                                        console.log(error)
+                                    })
                                 return true
                             })
                             .catch(error => {
                                 console.log(error)
                             })
                     } else {
-
+                        deleteAthleteData(userUID)
+                            .then(() => {
+                                admin.firestore()
+                                    .collection('users')
+                                    .doc(user.uid)
+                                    .delete()
+                                    .then(() => {
+                                        return true
+                                    })
+                                    .catch(error => {
+                                        console.log(error)
+                                    })
+                                return true
+                            })
+                            .catch(error => {
+                                console.log(error)
+                            })
                     }
                 }
-                return
+                return true
             })
-        //             admin.firestore()
-        //                 .collection('users')
-        //                 .doc(user.uid)
-        //                 .delete()
-        //                 .then(() => {
-        //                     return true
-        //                 })
-        //                 .catch(error => {
-        //                     console.log(error)
-        //                 })
+            .catch(error => {
+                console.log(error)
+            })
+
     })
+
+const deleteAthleteData = (docUID) => {
+    return new Promise((res, rej) => {
+        Promise.all([
+            deleteLocalExercises(docUID),
+            deletePrograms(docUID, 'athlete'),
+            deleteUserAssociations(docUID, 'athlete')
+        ])
+            .then(() => {
+                res(true)
+                return true
+            })
+            .catch(error => {
+                rej(error)
+            })
+    })
+}
 
 const deleteCoachData = (docUID) => {
     return new Promise((res, rej) => {
@@ -78,6 +116,7 @@ const deleteCoachData = (docUID) => {
         ])
             .then(() => {
                 res(true)
+                return true
             })
             .catch(error => {
                 rej(error)
@@ -86,11 +125,208 @@ const deleteCoachData = (docUID) => {
 }
 
 const deleteUserAssociations = (userUID, userType) => {
+    return new Promise((res, rej) => {
 
+        let targetUID = userType + "UID"
+
+        admin.firestore()
+            .collection('currentCoachAthletes')
+            .where(targetUID, '==', userUID)
+            .get()
+            .then(snap => {
+                if (!snap.empty) {
+                    const batch = admin.firestore().batch()
+                    let path = admin.firestore()
+                        .collection('currentCoachAthletes')
+
+                    snap.docs.forEach(doc => {
+                        batch.delete(path.doc(doc.id))
+                    })
+
+                    batch.commit()
+                        .then(() => {
+                            res(true)
+                            return true
+                        })
+                        .catch(error => {
+                            rej(error)
+                        })
+
+                } else {
+                    res(true)
+                }
+                return true
+            })
+            .catch(error => {
+                console.log(error)
+            })
+    })
 }
 
 const deletePrograms = (userUID, userType) => {
+    if (userType === 'coach') {
+        return new Promise((res, rej) => {
+            admin.firestore()
+                .collection('programs')
+                .where('owner', '==', userUID)
+                .where('status', '!=', 'past')
+                .get()
+                .then(snap => {
+                    if (!snap.empty) {
+                        const batch = admin.firestore().batch()
+                        let path = admin.firestore()
+                            .collection('programs')
 
+                        let promises = []
+
+                        snap.docs.forEach(doc => {
+                            promises.push(deleteProgramHelper(doc.id))
+                            batch.delete(path.doc(doc.id))
+                        })
+
+                        Promise.all(promises)
+                            .then(() => {
+                                batch.commit()
+                                    .then(() => {
+                                        res(true)
+                                        return true
+                                    })
+                                    .catch(error => {
+                                        rej(error)
+                                    })
+                                return true
+                            })
+                            .catch(error => {
+                                rej(error)
+                            })
+
+                    } else {
+                        res(true)
+                    }
+                    return true
+                })
+                .catch(error => {
+                    console.log(error)
+                })
+        })
+    } else {
+        return new Promise((res, rej) => {
+            admin.firestore()
+                .collection('programs')
+                .where('athlete', '==', userUID)
+                .get()
+                .then(snap => {
+                    if (!snap.empty) {
+                        const batch = admin.firestore().batch()
+                        let path = admin.firestore()
+                            .collection('programs')
+
+                        let promises = []
+                        snap.docs.forEach(doc => {
+                            promises.push(deleteProgramHelper(doc.id))
+                            batch.delete(path.doc(doc.id))
+                        })
+
+                        Promise.all(promises)
+                            .then(() => {
+                                batch.commit()
+                                    .then(() => {
+                                        res(true)
+                                        return true
+                                    })
+                                    .catch(error => {
+                                        rej(error)
+                                    })
+                                return true
+                            })
+                            .catch(error => {
+                                rej(error)
+                            })
+
+                    } else {
+                        res(true)
+                    }
+                    return true
+                })
+                .catch(error => {
+                    console.log(error)
+                })
+        })
+    }
+}
+
+const deleteProgramHelper = (docUID) => {
+    return new Promise((res, rej) => {
+        admin.firestore()
+            .collection('programs')
+            .doc(docUID)
+            .collection('exercises')
+            .get()
+            .then(snap => {
+                if (!snap.empty) {
+                    const batch = admin.firestore().batch()
+                    let path = admin.firestore()
+                        .collection('programs')
+                        .doc(docUID)
+                        .collection('exercises')
+
+                    snap.docs.forEach(doc => {
+                        batch.delete(path.doc(doc.id))
+                    })
+
+                    batch.commit()
+                        .then(() => {
+                            res(true)
+                            return true
+                        })
+                        .catch(error => {
+                            rej(error)
+                        })
+                } else {
+                    res(true)
+                }
+                return true
+            })
+            .catch(error => {
+                console.log(error)
+            })
+    })
+}
+
+const deleteLocalExercises = (userUID) => {
+    return new Promise((res, rej) => {
+        admin.firestore()
+            .collection('exercises')
+            .where('owner', '==', userUID)
+            .get()
+            .then(snap => {
+                if (!snap.empty) {
+                    const batch = admin.firestore().batch()
+                    let path = admin.firestore()
+                        .collection('exercises')
+
+                    snap.docs.forEach(doc => {
+                        batch.delete(path.doc(doc.id))
+                    })
+
+                    batch.commit()
+                        .then(() => {
+                            res(true)
+                            return true
+                        })
+                        .catch(error => {
+                            rej(error)
+                        })
+
+                } else {
+                    res(true)
+                }
+                return true
+            })
+            .catch(error => {
+                console.log(error)
+            })
+    })
 }
 
 const deleteCoachTeams = (docUID) => {
@@ -99,8 +335,32 @@ const deleteCoachTeams = (docUID) => {
         admin.firestore()
             .collection('users')
             .doc(docUID)
-            .delete()
-            .then(() => {
+            .collection('teams')
+            .get()
+            .then(snap => {
+                if (!snap.empty) {
+                    const batch = admin.firestore().batch()
+                    let path = admin.firestore()
+                        .collection('users')
+                        .doc(docUID)
+                        .collection('teams')
+
+                    snap.docs.forEach(doc => {
+                        batch.delete(path.doc(doc.id))
+                    })
+
+                    batch.commit()
+                        .then(() => {
+                            res(true)
+                            return true
+                        })
+                        .catch(error => {
+                            rej(error)
+                        })
+
+                } else {
+                    res(true)
+                }
                 return true
             })
             .catch(error => {
@@ -114,12 +374,30 @@ const deleteCoachProgramGroups = (docUID) => {
         admin.firestore()
             .collection('users')
             .doc(docUID)
-            .delete()
-            .then(() => {
+            .collection('programGroups')
+            .get()
+            .then(snap => {
+                if (!snap.empty) {
+                    admin.firestore()
+                        .collection('users')
+                        .doc(docUID)
+                        .collection('programGroups')
+                        .doc('programGroups')
+                        .delete()
+                        .then(() => {
+                            res(true)
+                            return true
+                        })
+                        .catch(error => {
+                            rej(error)
+                        })
+                } else {
+                    res(true)
+                }
                 return true
             })
             .catch(error => {
-                console.log(error)
+                rej(error)
             })
     })
 }
